backend/requirements.txt
flask
flask-cors
openai
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23